pyproject.toml
[tool.poetry]
package-mode = false
name = "pokerbot"
version = "0.1.0"
description = ""
authors = ["Oleksandr Nazarevych <oleksandr.o.nazarevych@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
mss = "^10.0.0"
opencv-python = "^4.11.0.86"
pytesseract = "^0.3.13"
pydantic = "^2.10.6"
requests = "^2.32.3"
bs4 = "^0.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
